+++ APP/msg/app_msg_def.h
@@ -7,7 +7,7 @@
 #include <cstdint>
 
 struct app_msg_gimbal_to_chassis {
-    float vx, vy, rotate;
+    int16_t vx, vy, rotate;
     int8_t save_state[2];
 
 } __attribute__((packed));
